修改管理员权限

+++ price/php/product/allproducts.php
@@ -96,7 +96,8 @@ require_once '../../util/functions.php';
 <div id="content-outer">
     <!-- start content -->
     <div id="content">
-        <h1 style="color: #1a1a1a; padding: 10px 0 20px 20px; font-size: 25px">所有商品信息</h1><h3><a href="deleteAllProduct.php">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;删除所有商品（谨慎操作！）</a></h3>
+        <h1 style="color: #1a1a1a; padding: 10px 0 20px 20px; font-size: 25px">所有商品信息</h1><h3><?php $user_id_admin=$_SESSION['id'];
+            if ($user_id_admin==1||$user_id_admin==2){ echo '<a href="deleteAllProduct.php">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;删除所有商品（谨慎操作！）</a>';} ?></h3>
         <table border="0" width="100%" cellpadding="0" cellspacing="0" id="content-table">
             <tr>
                 <td></td>
@@ -111,7 +112,6 @@ require_once '../../util/functions.php';
                 <th rowspan="3" class="sized"><img src="../../images/shared/side_shadowright.jpg" width="20"
                                                    height="300"
                                                    alt=""/></th>
-
             </tr>
             <tr>
                 <td id="tbl-border-left"></td>
@@ -135,39 +135,44 @@ require_once '../../util/functions.php';
                                         <th class="table-header-repeat line-left minwidth-1">监控开关</th>
                                     </tr>
                                     <?php
-                                    connectDB();
-                                    $result = mysql_query("SELECT * FROM monitor ");
-                                    $data_count = mysql_num_rows($result);
-                                    for ($i = 0; $i < $data_count; $i++) {
-                                        $result_arr = mysql_fetch_assoc($result);
-                                        $id = $result_arr['id'];
-                                        $item_id = $result_arr['item_id'];
-                                        $item_name = $result_arr['item_name'];
-                                        $mall_name = $result_arr['mall_name'];
-                                        if ($mall_name=='jd'){
-                                            $mall_name_ven='京东';
-                                        }elseif ($mall_name=='tb'){
-                                            $mall_name_ven='淘宝';
-                                        }elseif($mall_name=='tm'){
-                                            $mall_name_ven='天猫';
-                                        }else{
-                                            $mall_name_ven=$mall_name;
-                                        }
-                                        $item_price = $result_arr['item_price'];
-                                        $user_id = $result_arr['user_id'];
-                                        $add_date = $result_arr['add_date'];
-                                        $user_result = mysql_query("SELECT * FROM user WHERE user_id = $user_id");
-                                        $user_result_arr = mysql_fetch_assoc($user_result);
-                                        $user_name = $user_result_arr['user_name'];
-                                        $status = $result_arr['status'];
-                                        $note = $result_arr['note'];
-                                        $status == 0 ? ($statusCode = '<span style="color: red">尚未监控</span>') : ($statusCode = '<span style="color: green">正在监控</span>');
-                                        $user_price = $result_arr['user_price'];
-                                        echo "<tr><td>$item_id</td><td><a href='https://item.jd.com/$item_id.html'>$item_name</a></td><td>$mall_name_ven</td><td>$item_price</td><td>$user_price</td><td>$statusCode</td>
+                                    if ($user_id_admin==1||$user_id_admin==2){
+                                        connectDB();
+                                        $result = mysql_query("SELECT * FROM monitor ");
+                                        $data_count = mysql_num_rows($result);
+                                        for ($i = 0; $i < $data_count; $i++) {
+                                            $result_arr = mysql_fetch_assoc($result);
+                                            $id = $result_arr['id'];
+                                            $item_id = $result_arr['item_id'];
+                                            $item_name = $result_arr['item_name'];
+                                            $mall_name = $result_arr['mall_name'];
+                                            if ($mall_name=='jd'){
+                                                $mall_name_ven='京东';
+                                            }elseif ($mall_name=='tb'){
+                                                $mall_name_ven='淘宝';
+                                            }elseif($mall_name=='tm'){
+                                                $mall_name_ven='天猫';
+                                            }else{
+                                                $mall_name_ven=$mall_name;
+                                            }
+                                            $item_price = $result_arr['item_price'];
+                                            $user_id = $result_arr['user_id'];
+                                            $add_date = $result_arr['add_date'];
+                                            $user_result = mysql_query("SELECT * FROM user WHERE user_id = $user_id");
+                                            $user_result_arr = mysql_fetch_assoc($user_result);
+                                            $user_name = $user_result_arr['user_name'];
+                                            $status = $result_arr['status'];
+                                            $note = $result_arr['note'];
+                                            $status == 0 ? ($statusCode = '<span style="color: red">尚未监控</span>') : ($statusCode = '<span style="color: green">正在监控</span>');
+                                            $user_price = $result_arr['user_price'];
+                                            echo "<tr><td>$item_id</td><td><a href='https://item.jd.com/$item_id.html'>$item_name</a></td><td>$mall_name_ven</td><td>$item_price</td><td>$user_price</td><td>$statusCode</td>
                                               <td>$user_name</td><td>$add_date</td><td><a href='editProduct.php?id=$id&&edit_auth=1'>修改</a>
                                               |<a href='deleteProduct.php?id=$id&&delete_auth=1'>删除</a></td>
                                               <td><a href=\"../product/monitorSwitch.php?status=1&&id=$id&&auth=1\" class=\"icon-5 info-tooltip\"></a><a href=\"../product/monitorSwitch.php?status=0&&id=$id&&auth=1\" class=\"icon-2 info-tooltip\"></a></td></tr>";
+                                        }
+                                    }else{
+                                        echo '<tr><td colspan="10" style="color: red;font-size: 20px;text-align: center"> 你的访问权限不够,如果有疑问，请联系管理员！</td></tr>';
                                     }
+
                                     ?>
                                 </table>
                                 <!--  end product-table................................... -->
@@ -206,7 +211,7 @@ require_once '../../util/functions.php';
 <div id="footer">
     <!--  start footer-left -->
     <div id="footer-left">
-        &copy; Copyright BY ShaTong. <span id="spanYear"></span> <a href="http://jd.usau-buy.me/"></a>. All rights
+        &copy; Copyright BY ShaTong. <span id="spanYear"></span> <a href="http://monitor.usau-buy.me/"></a>. All rights
         reserved.
     </div>
     <!--  end footer-left -->
